better error reporting in hmap2, now prints line number from .map file for many errors and warnings Entity with no valid brushes error is now only a warning
git-svn-id: svn://svn.icculus.org/twilight/trunk@10391 d7cf8633-e32d-0410-b094-e92efae38249

+++ hmap2/light_face.c
@@ -186,7 +186,7 @@ static void CalcSamples( lightinfo_t *l )
 	int				mapnum;
 
 	l->numsamples = l->texsize[0] * l->texsize[1];
-	if( l->numsamples > SINGLEMAP )
+	if( (unsigned int)l->numsamples > (unsigned int)SINGLEMAP )
 		Error( "Bad lightmap size: %i", l->numsamples );
 
 	if( l->numsamples > l->maxsamples ) {
